+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 #### 1.9.6-SNAPSHOT
 
 - **[UPDATE]** Updated to Kotlin 1.1.3-2.
+- **[UPDATE]** Updated to Kotlin Coroutines 0.17.
+- **[UPDATE]** Updated to Gradle 4.0.2.
 - **[FEATURE]** (`ktx-box2d`) `World.rayCast` extension methods that allow creating ray-cast callbacks with the Kotlin
 lambda syntax. `KtxRayCastCallback` alias added to ease implementation of this utility.
 - **[FEATURE]** (`ktx-box2d`) Added `RayCast` object with constants that can be returned by the custom `RayCastCallback` implementations.